remove cookies method from httpmessage trait
